requirements.txt
langchain==0.3.26
langchain-community==0.3.27
langchain-openai==0.3.27
langchain-core==0.3.75
langchain-aws==0.2.1
langchain-google-vertexai==2.0.27
langchain_text_splitters==0.3.8 # 0.3.3
boto3==1.34.144
sqlalchemy==2.0.41
python-dotenv==1.1.1
fastapi==0.115.12
psycopg2-binary==2.9.9
pgvector==0.2.5
uvicorn==0.28.0
pypdf==6.0.0
unstructured==0.16.11
markdown==3.8.2
networkx==3.2.1
pandas==2.2.1
openpyxl==3.1.5
docx2txt==0.9
pypandoc==1.15
PyJWT==2.8.0
asyncpg==0.29.0
python-multipart==0.0.19
sentence_transformers==3.1.1
aiofiles==24.1.0
rapidocr-onnxruntime==1.4.4
opencv-python-headless==4.9.0.80
pymongo==4.6.3
langchain-mongodb==0.2.0
langchain-ollama==0.3.3
langchain-huggingface==0.1.0
langchain-google-genai==2.1.10
cryptography==45.0.5
python-magic==0.4.27
python-pptx==1.0.2
xlrd==2.0.2
pydantic==2.9.2
chardet==5.2.0
promptfoo
